Add a few positive tests

+++ osu.Framework/Bindables/BindableList.cs
@@ -234,6 +234,9 @@ namespace osu.Framework.Bindables
         {
             ensureMutationAllowed();
 
+            if (count == 0)
+                return;
+
             var removedItems = collection.GetRange(index, count);
 
             collection.RemoveRange(index, count);
